requirements/requirements.txt
pyautogui
wheel
psutil
pywin32
keyboard
pillow
opencv-python